# apps/backend/gen/swagger/swagger.yaml
definitions:
  response.ErrorResponse:
    properties:
      details:
        items:
          type: string
        type: array
      message:
        type: string
    type: object
  video.CreateVideoRequest:
    properties:
      description:
        maxLength: 1000
        type: string
      title:
        maxLength: 100
        minLength: 1
        type: string
      visibility:
        $ref: '#/definitions/video.Visibility'
    required:
    - title
    - visibility
    type: object
  video.Response:
    properties:
      description:
        type: string
      duration_ms:
        type: integer
      file_path:
        type: string
      file_size:
        type: integer
      id:
        type: string
      mime_type:
        type: string
      owner_id:
        type: string
      status:
        type: string
      title:
        type: string
      visibility:
        type: string
    type: object
  video.UpdateVideoRequest:
    properties:
      description:
        maxLength: 1000
        type: string
      title:
        maxLength: 100
        minLength: 1
        type: string
      visibility:
        $ref: '#/definitions/video.Visibility'
    type: object
  video.Visibility:
    enum:
    - private
    - public
    type: string
    x-enum-varnames:
    - Private
    - Public
info:
  contact: {}
  description: Route Streaming Service Backend API.
  title: Route Streaming API.
  version: 1.0.
paths:
  /health:
    get:
      description: サーバーの稼働状態を確認します
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: ヘルスチェック
      tags:
      - health
  /videos:
    get:
      consumes:
      - application/json
      description: 登録されている動画の一覧を取得します
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/video.Response'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: 動画一覧取得
      tags:
      - videos
    post:
      consumes:
      - application/json
      description: 新しい動画のメタデータを作成します
      parameters:
      - description: Create Route Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/video.CreateVideoRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/video.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: 動画新規作成
      tags:
      - videos
  /videos/{id}:
    delete:
      consumes:
      - application/json
      description: 指定したIDの動画を削除します
      parameters:
      - description: Route ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: 動画削除
      tags:
      - videos
    get:
      consumes:
      - application/json
      description: IDを指定して動画の詳細情報を取得します
      parameters:
      - description: Route ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/video.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: 動画詳細取得
      tags:
      - videos
    put:
      consumes:
      - application/json
      description: 指定したIDの動画メタデータを更新します
      parameters:
      - description: Route ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Route Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/video.UpdateVideoRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: 動画更新
      tags:
      - videos
  /videos/{id}/stream:
    get:
      description: 指定したIDの動画をHTTP Range対応ストリーミング配信します
      parameters:
      - description: Video ID
        in: path
        name: id
        required: true
        type: string
      - description: Range Header
        in: header
        name: Range
        type: string
      produces:
      - video/mp4
      responses:
        "200":
          description: OK
          schema:
            type: file
        "206":
          description: Partial Content
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: 動画ストリーミング配信
      tags:
      - stream
swagger: "2.0"